Temp File
This commit adds a temp directory that incases files that can me modified for future use with the contact form. This as well has updated the index.php to have lorem ipsum as a place holder until content is given. Added styles.css for all css other than the nav.css.