Css/Index updates
This updates the styles.css and index.php files. Index.php has the change on line 35 to be an aside vs an article. The styles .css has changes to the h2 colors, adds in article#gamingPage styles as well as img.crown and img.legacy styles. Changes the BG and text color in the footer section. Adds in aside.reservation styles.