Header updates
This commit adds a clear: both to the nav.css on line 33 which is part of the header issue.  This as well changes a few things on the header.php file this changers the link for the css from css/styles.css to ../css/styles.css. Ads in the header tag and moves the body opening tag to line 36.