Index/contact/styles update
Index.php update to remove as < that was on the page that was not needed.  Contact.php updated to have the class on the aside be leftColumn instead of rightcolumn. Added an <h2> tag on line 4 with a class of contactUs. Styles.css updted for the font-size on line 33 added in aside.reservations styles, added in aside.leftColumn styles and h2.contactUs styles.