Forms
added forms to index.php and contact.php. Updated form.css to account for the form styles. Updated Nav.css to remove the postion:fixed and width. updated header to have the new link in the list under storeInfo. Updated storeinfo.php to include the Turn Back Time Zone for the clients store.