index and storeinfo updates
Added titles to each of the sections for the form on the home page. Added a new addtion to the reservations form on the home page for Time. Took out the placeholder text for the phases and added in title of h3 for the phases. Started adding in the phase one text.