File updates
Added in a new file of readMore.php, updated index.php to have the h1 outside of the div. Changed the <a href=#> to readMore.php. Updated classes in each section for hte <p> or the <div>. Added in a section for gamingAtFinest and retroInfo. Updated CSS to layout the page.