Added in Tablet Styles
Updated index.php and header.php with classes for desktop and tablet. Moved some images in index.php to be above where they were previously on lines 77 and 78 to lines 64 to 66 and did the same for lines 80 and 81 from 92 to 94. Tablet has been added in max witdh is 1331px, min width is 959px.